+++ src/platform_impl/web/mod.rs
@@ -43,5 +43,5 @@ pub(crate) use self::keyboard::KeyEventExtra;
 pub(crate) use crate::icon::NoIcon as PlatformIcon;
 pub(crate) use crate::platform_impl::Fullscreen;
 pub(crate) use cursor::CustomCursor as PlatformCustomCursor;
-pub(crate) use cursor::CustomCursorBuilder as PlatformCustomCursorBuilder;
 pub(crate) use cursor::CustomCursorFuture;
+pub(crate) use cursor::CustomCursorSource as PlatformCustomCursorSource;